Hasing, Logging and Spring Data

+++ src/main/java/nc/Medas/repo/UserRepo.java
@@ -2,12 +2,14 @@ package nc.Medas.repo;
 
 import nc.Medas.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Repository;
 
 import java.math.BigInteger;
 
 @Repository
-public interface UserRepo  extends JpaRepository<User, Long> {
+public interface UserRepo  extends JpaRepository<User, Integer> {
     User findByLogin(String login);
 
 }
